[refactor] Move sign controller file from pkg to signctl directory
