+++ pkg/registry/registry.go
@@ -1,7 +1,6 @@
 package registry
 
 import (
-	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/json"
@@ -14,9 +13,6 @@ import (
 	"time"
 
 	"github.com/docker/distribution/registry/client/auth/challenge"
-	apiv1 "github.com/tmax-cloud/registry-operator/api/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	regClient "github.com/docker/distribution/registry/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -32,50 +28,6 @@ type Repository struct {
 }
 
 var logger = logf.Log.WithName("pkg-registry-api")
-
-func getRegistry(c client.Client, regName, namespace string) (*apiv1.Registry, error) {
-	reg := &apiv1.Registry{}
-	if err := c.Get(context.TODO(), types.NamespacedName{Name: regName, Namespace: namespace}, reg); err != nil {
-		return nil, err
-	}
-
-	return reg, nil
-}
-
-type RegCtl struct {
-	client client.Client
-	reg    *apiv1.Registry
-}
-
-// NewRegCtl is a controller for registry
-// if registryName or registryNamespace is empty string, RegCtl is nil
-func NewRegCtl(c client.Client, regName, namespace string) *RegCtl {
-	if len(regName) == 0 || len(namespace) == 0 {
-		return nil
-	}
-
-	reg, err := getRegistry(c, regName, namespace)
-	if err != nil {
-		return nil
-	}
-
-	return &RegCtl{
-		client: c,
-		reg:    reg,
-	}
-}
-
-func (r *RegCtl) GetHostname() string {
-	return strings.TrimPrefix(r.GetEndpoint(), "https://")
-}
-
-func (r *RegCtl) GetEndpoint() string {
-	return r.reg.Status.ServerURL
-}
-
-func (r *RegCtl) GetNotaryEndpoint() string {
-	return r.reg.Status.NotaryURL
-}
 
 const DefaultServer = "https://registry-1.docker.io"
